Add comprehensive Phase 2 Hessian analysis certification directory
- Create ForwardDiff_Certification/ with complete test suite and documentation
- Add eigenvalue-based critical point classification using ForwardDiff.jl
- Include comprehensive unit tests, demos, and visualization examples
- Add Phase 2 implementation files: hessian_analysis.jl, hessian_visualization.jl
- Create detailed certification documentation and validation results
- Add test files for integration and workflow validation

+++ src/LibFunctions.jl
@@ -400,7 +400,7 @@ end
 
 end
 
-@doc nothing function Rastringin(x::AbstractVector)::Float64
+@doc nothing function Rastringin(x::AbstractVector)
     # =======================================================
     #   Not Rescaled
     #   Rastringin function
